Feature/modificaciones service

+++ src/main/java/com/novabank/dto/CuentaDTO.java
@@ -2,6 +2,7 @@ package com.novabank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,6 +10,6 @@ public class CuentaDTO {
 
     private Long id;
     private String numeroCuenta;
-    private Double saldo;
+    private BigDecimal saldo;
     private LocalDateTime fechaCreacion;
 }
